Front terminé, ne reste que le bug du CORS policy quand on passe par la gateway

+++ Service-Commande-Article/src/main/java/fr/uphf/PABLCommandeArticle/dto/PostCommandeRequest.java
@@ -19,6 +19,6 @@ public class PostCommandeRequest {
     private String details;
     private Integer idRestaurant;
     private String status;
-    private List<Long> articles;
+    private List<Integer> articles;
 }
